Refactor to workflow page
