+++ src/app.html
@@ -1,17 +1,19 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <link rel="icon" href="/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="stylesheet" href="/css/normalize.css" />
-    <link rel="stylesheet" href="/css/override.css" />
-    <link rel="stylesheet" href="/prism/prism.css" />
-    %svelte.head%
-  </head>
 
-  <body>
-    <div id="svelte">%svelte.body%</div>
-    <script src="/prism/prism.js"></script>
-  </body>
+<head>
+  <meta charset="utf-8" />
+  <link rel="icon" href="/favicon.ico" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <link rel="stylesheet" href="/css/normalize.css" />
+  <link rel="stylesheet" href="/css/override.css" />
+  <link rel="stylesheet" href="/prism/prism.css" />
+  %sveltekit.head%
+</head>
+
+<body>
+  <div id="svelte">%sveltekit.body%</div>
+  <script src="/prism/prism.js"></script>
+</body>
+
 </html>
