+++ src/lib/utilities/query/to-list-workflow-parameters.ts
@@ -8,10 +8,10 @@ type ParsedParameters = FilterParameters & { timeRange?: string };
 
 const is =
   (identifier: string) =>
-  (token: string): boolean => {
-    if (token.toLowerCase() === identifier.toLowerCase()) return true;
-    return false;
-  };
+    (token: string): boolean => {
+      if (token.toLowerCase() === identifier.toLowerCase()) return true;
+      return false;
+    };
 
 const getTwoAhead = (tokens: Tokens, index: number): string =>
   tokens[index + 2];
